% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__networkAnalyser__alleleClusterDistr.R
\name{alleleClusterDistr}
\alias{alleleClusterDistr}
\title{Compute a matrix for presence of allele clusters based on allelic presence-absence status}
\usage{
alleleClusterDistr(cls, pam)
}
\arguments{
\item{cls}{A data frame of two columns for cluster IDs (first column) and allele
contents (the second column, comma-delimited). This data frame can be generated
from the output of function summariseCliques. Each cluster must consist of at
least two alleles.}

\item{pam}{An uncentred presence-absence matrix for alleles in the data frame cls.
This is a binary matrix, where 1 denotes presence and 0 denotes absence.}
}
\value{
An n-by-m matrix, where n is the sample size and m is the cluster number.
}
\description{
This function generates a matrix for co-occurrence of all alleles
of each cluster in a give collection of bacterial samples. This function is
particularly useful in investigating the distribution of allele clusters (such
as maximal cliques) in bacterial samples. It can be considered as a simplified
version of function getClusterMemberCoocurrence.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
