% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__networkAnalyser__allelicCoMatrix.R
\name{allelicCoMatrix}
\alias{allelicCoMatrix}
\title{Create a co-occurrence matrix from allelic PAM}
\usage{
allelicCoMatrix(
  clusters = NULL,
  pam = NULL,
  hclust.method = NULL,
  hclust.dist = "binary"
)
}
\arguments{
\item{clusters}{A named or unnamed list of allele name vectors that define
alleles of each cluster. Every vector must contain at least two alleles.}

\item{pam}{An allelic presence-absence matrix, where 1 denotes presence of an
allele and 0 denotes absence of an allele.}

\item{hclust.method}{Method for clustering columns in the output co-occurrence
matrix. Default: NULL (no clustering)}

\item{hclust.dist}{Distance metric used for column-wise clustering. Default:
binary.}
}
\value{
A list of elements: m, a co-occurrence matrix; c, clusters, recordered
if hclust.method != NULL; g: a vector (method, dist) for arguments of
column-wise clustering (grouping).
}
\description{
Create a co-occurrence matrix from allelic PAM
}
\note{
hclust.method and hclust.dist are passed to R base functions hclust and
dist as arguments.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
