% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__assignPairID.R
\name{assignPairID}
\alias{assignPairID}
\title{Assign an ID to each pair of alleles or patterns.}
\usage{
assignPairID(lmms, from = 1, paired.rows = TRUE)
}
\arguments{
\item{lmms}{A data frame whose first two columns (named y and x respectively) are used to assign pair IDs.}

\item{from}{The initial pair ID.}

\item{paired.rows:}{whether there are only two rows per pair of alleles, each in an inverse direction.
Set paired.rows = FALSE to process allele pairs in the data frame lmms.ds[["dif"/"idd"]].}
}
\description{
This is a generic function that assigns an ID to each pair of object names. For example, a single ID is
assigned to two rows in a data frame for associations between patterns (pat_1, pat_2) and (pat_2, pat_1).
}
\examples{
a0 <- assoc[["lmms.pat"]][["dif"]][["h1"]]
a0 <- assignPairID(lmms = a0, from = 1, paired.rows = TRUE)

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
