% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__betaSignCongruence.R
\name{betaSignCongruence}
\alias{betaSignCongruence}
\title{Calculate sign congruence of fixed effects (betas) between linear mixed
models (LMMs) and penalised logistic models (PLMs)}
\usage{
betaSignCongruence(lmms, plms, p.max = 1, step = 5e-04, nevLog = TRUE)
}
\arguments{
\item{lmms}{A data frame from the function lmm for association status between
patterns. Assuming assoc = lmm(...), then lmms = assoc$lmms.pat$dif$h1.}

\item{plms}{A data frame from the function plr for association status between
patterns. Assuming assoc = plr(...), then plms = assoc$pat. Pattern orders for
association tests must match to those in lmms. Otherwise, both data frames
cannot merge correctly.}

\item{p.max}{Upper bound of raw p-values for congruence assessment.}

\item{step}{Step size of betas for calculating the congruence percentage.}

\item{nevLog}{A logical argument determining whether -log10 will be applied to
p-values in the result.}
}
\value{
A data frame of percentages as measurements of sign congruence.
}
\description{
This function calculates the percentage of matched fixed effects
having the same sign. The outcome can be visualised as a scatter plot or so.
}
\examples{
sg <- betaSignCongruence(lmms = assoc_lmm$lmms.pat$dif$h1,
plms = assoc_lgr$pat, p.max = 1, step = 0.0005, nevLog = TRUE)

}
\author{
Yu Wan (\email{wanyuac@126.com})
}
