% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/func__geoTempAnalyser__calcAllelicDiversityPerCountry.R
\name{calcAllelicDiveristyPerCountry}
\alias{calcAllelicDiveristyPerCountry}
\title{Allelic diversity per country or any kind of regions}
\usage{
calcAllelicDiveristyPerCountry(sam, pam, alleles = NULL,
  countries = NULL, method = "simpson", shannon_base = exp(1))
}
\arguments{
\item{sam}{A data frame whose first and second columns are strain and country
names, respectively.}

\item{pam}{A binary allelic presence-absence matrix, where row names are
strain names and column names are allele names.}

\item{alleles}{A character vector specifying alleles to be analysed. Keep NULL
to include all alleles in pam.}

\item{countries}{A character vector specifying countries to be considered. Keep
NULL to include all countries in pam.}

\item{method}{A character argument of either "simpson" or "shannon".}

\item{shannon_base}{The base for Shannon's diversity index. It is not used for
Simpson's diveristy index. Default: exp(1).}
}
\value{
A list of four elements:
  c: a matrix of n(countries) by n(alleles) for allele counts per country;
  d: a named numeric vector of diversity indices.
  p: the final allelic presence-absence matrix on which the diversity indices
are computed.
  s: isolation information of the final set of strains.
}
\description{
Calculate Simpson's or Shannon's diversity index for alleles per
country or any other kind of geographic regions. Each strain must belong to
only a single country.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
