% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__networkAnalyser__compEdgeOccur.R
\name{compEdgeOccur}
\alias{compEdgeOccur}
\title{Compare similarity between edges' occurrence}
\usage{
compEdgeOccur(nwk, apam, directed = TRUE)
}
\arguments{
\item{nwk}{A Graph object created by the function mkNetwork for a single network.
It has three attributes: id, E and V. Required columns of E are "y" and "x",
which are vertex names.}

\item{apam}{Allelic presence-absence matrix. Rows: strain names; columns:
allele names.}

\item{directed}{A logical argument specifying whether nwk stores a directed network.
Default: TRUE.}
}
\value{
A list of three elements: nwk, an undirected network converted from
the input network; ep, a data frame of edge pairs; sim, the similarity matrix
for edges, which is a symmetric matrix.
}
\description{
Compute a matrix of Simpson's similarity coefficients for the
presence of edges.
}
\examples{
compEdgeOccur

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
