% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__networkAnalyser__compileGraphs.R
\name{compileGraphs}
\alias{compileGraphs}
\title{Compile graphs in a Graph object into a single network while retaining
separation between individual graphs.}
\usage{
compileGraphs(graphs, edge.attr = NULL, node.attr = NULL)
}
\arguments{
\item{graphs}{A Graph object created using the function extractSubgraphs.}

\item{edge.attr}{A character vector of column names for edge attributes.}

\item{node.attr}{A character vector of column names for node attributes.}
}
\description{
This function pools graphs in a Graph object into a single network
while keeping individual graphs separated. It is particularly useful when a
user wants to display cliques or other kinds of clusters in a single view (such
as in Cytoscape). The Graph object is produced by the function extractSubgraphs.

This function does not behave like the Cytoscape utility "merge networks".
Specifically, it keeps the separation between graphs through appending a cluster
ID to each node name so that the node shared by different graphs become distinguishable.
It adds a label column to save the node names. In Cytoscape, for example, the
user can choose this label column as node labels.
}
\examples{
g <- compileGraphs(graphs = clusters, edge.attr = c("beta", "score"),
node.attr = c("n", "class"))

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
