% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__visualisation__compPvalues.R
\name{compPvalues}
\alias{compPvalues}
\title{Draw a scatter plot and histograms to compare p-values from linear mixed models
(LMMs) and penalised logistic models (PLMs).}
\usage{
compPvalues(p.lmm, p.plm, lmm.h0, p.min = 2.2e-16, p.max = 0.05,
  L.weak = 1, bks = seq(0, 16, by = 2), show.p.max = TRUE, cols = c(`0`
  = "red", `1` = "orange", `2` = "blue", `3` = "grey50"),
  plot.title.size = 10, axis.text.size = 8, axis.title.size = 10,
  axis.title.size.panelB = 3, img = "comp_pvals.png", img.w = 150,
  img.h = 150, img.r = 300, img.u = "mm", img.oma = c(0.1, 0.1, 0.1,
  0.1), img.mar = c(3, 2.6, 1, 0.5), img.mgp = c(1.8, 0.6, -0.2),
  p.prev = NULL)
}
\arguments{
\item{p.lmm}{A data frame from the function lmm for association status between
patterns. Assuming assoc = lmm(...), then p.lmm = assoc$lmms.pat$dif$h1.}

\item{p.plm}{A data frame from the function plr for association status between
patterns. Assuming assoc = plr(...), then p.lmm = assoc$pat. Pattern orders for
association tests must match to those in p.lmm. Otherwise, both data frames
cannot merge correctly.}

\item{lmm.h0}{A data frame from the function lmm for null models. lmm.h0 =
assoc$lmms.pat$dif$h0.}

\item{p.min}{Minimum of raw p-values that can be represented in computers
precisely. Default: 2.2e-16.}

\item{p.max}{Maximum of Bonferroni-corrected p-values for concluding
significance. Default: 0.05.}

\item{L.weak}{Maximum log10 lambda for defining weak structural random effects.
Default: 1.}

\item{bks}{Breaks for X and Y axes in the plot. Default: 0, 2, 4, 6, ..., 16.}

\item{show.p.max}{A logical argument determine whether to draw a grey dashed
line on each axis to indicate the cutoff for significance. Default: TRUE.}

\item{cols}{A character vector with names "0", "1", "2" and "3" for definition
of colours for lambda groups.}

\item{plot.title.size}{Size of the title of each panel.}

\item{axis.text.size}{Size of axis labels.}

\item{axis.title.size}{Size of axis titles.}

\item{axis.title.size.panelB}{Title sizes for panel b, which should be much
smaller than axis.title.size and controls the alignment of panels.}

\item{img}{Name and path for the output PNG file. Default: comp_pvals.png.}

\item{img.w}{Image width.}

\item{img.h}{Image height.}

\item{img.r}{Image resolution.}

\item{img.u}{Unit for the image width and height.}

\item{img.oma}{The argument oma for the function par during plotting.}

\item{img.mar}{The argument mar for the function par.}

\item{img.mgp}{The argument mgp for the function par.}

\item{p.prev}{Output of this function from a previous run. It overrides p.lmm
and p.plm.}
}
\value{
A data frame underlying the output plot.
}
\description{
This function only works at the pattern level. P-values are grouped
by log10 lambda0 of X and Y.
Dependencies: ggplot2, grid and gridExtra.
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
