% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__corPatternProj.R
\name{corPatternProj}
\alias{corPatternProj}
\title{Estimating correlation between an allelic distribution pattern and each
projection vector.}
\usage{
corPatternProj(pat.mat, proj.mat)
}
\arguments{
\item{pat.mat}{A uncentred pattern matrix, which is the element alleles$B in
the output of the function lmm or findPhysLink.}

\item{proj.mat}{A projection matrix, which is the element struc$C$C in the
output of the function lmm or findPhysLink.}
}
\description{
This function computes a point-biserial correlation coefficient
between each pattern and each projection vector. It works in a similar way as
the function corCladeProj. The point-biserial correlation coefficient is a
special case of the well-known Pearson correlation coefficient.
}
\examples{
a <- lmm(...)
r <- corPatternProj(pat.mat = a$alleles$B, proj.mat = a$struc$C$C)

}
\author{
Yu Wan (\email{wanyuac@126.com})
}
