% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__countAlleles.R
\name{countAlleles}
\alias{countAlleles}
\title{Mapping alleles to genes}
\usage{
countAlleles(pam.genes, pam.alleles, patterns = NULL, pat.sizes = NULL)
}
\arguments{
\item{pam.genes}{a compiled matrix showing allele calls across isolates, where extended allele IDs have been assigned}

\item{pam.alleles}{a presence-absence matrix of alleles, which is derived from the matrix pam.genes.}

\item{patterns}{(optional) a data frame consisting of two columns: allele and pattern ID}

\item{pat.sizes}{(optional) a two-column data frame (pattern and size) for the number of alleles per pattern}
}
\description{
Make a data frame to summarise genes, alleles and allele frequencies.
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
