% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__geoTempAnalyser__countAllelesPerCountry.R
\name{countAllelesPerCountry}
\alias{countAllelesPerCountry}
\title{Count alleles per country}
\usage{
countAllelesPerCountry(alleles = NULL, sam, mapping, apam, nul = NA,
  freq = FALSE)
}
\arguments{
\item{alleles}{A character vector of allele names.}

\item{sam}{A data frame of sample information. Required columns: Strain, Country, Year_low and Year_up.}

\item{mapping}{A data frame in the output of lmm or findPhysLink.}

\item{apam}{Allelic presence-absence matrix.}

\item{nul}{A value for zero counts. It can be zero definitely, but setting it to NA makes it easier to draw a heat map.}

\item{freq}{A logical argument determining if the allele frequency (in %)
rather than allele count is returned.}
}
\value{
This function returns a list of two elements: count and mapping, when
freq = FALSE, else, returns a list of three elements: freq, mapping and sample_size_country.
}
\description{
Count alleles per country
}
\examples{
a_nat <- countAllelesPerCountry(alleles = nwk$V$allele, sam = sam, mapping = assoc$mapping, apam = assoc$alleles$A, nul = -30)

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
