% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/func__geoTempAnalyser__countAllelesPerGeneByYear.R
\name{countAllelesPerGeneByYear}
\alias{countAllelesPerGeneByYear}
\title{Count each gene per year using their alleles}
\usage{
countAllelesPerGeneByYear(sam, mapping, gm, nul = NA)
}
\arguments{
\item{sam}{A data frame of sample information. Required columns: Strain, Country, Year_low and Year_up.}

\item{mapping}{A data frame in the output of lmm or findPhysLink.}

\item{gm}{Gene content matrix. Its column names must match the gene names in the mapping data frame.}

\item{nul}{A value for zero counts. It can be zero definitely, but setting it to NA makes it easier to draw a heat map.}
}
\description{
Neither function cares any alleles.
}
\examples{
ac <- countAllelesPerGeneByYear(sam = sam, mapping = assoc$mapping, gm = assoc$genes$all, nul = -30)

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
