% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__geoTempAnalyser__countAllelesPerYear.R
\name{countAllelesPerYear}
\alias{countAllelesPerYear}
\title{Count alleles per year}
\usage{
countAllelesPerYear(alleles = NULL, sam, mapping, apam, nul = NA)
}
\arguments{
\item{alleles}{A character vector of allele names.}

\item{sam}{A data frame of sample information. Required columns: Strain, Country, Year_low and Year_up.}

\item{mapping}{A data frame in the output of lmm or findPhysLink.}

\item{apam}{Allelic presence-absence matrix.}

\item{nul}{A value for zero counts. It can be zero definitely, but setting it to NA makes it easier to draw a heat map.}
}
\examples{
a_yr <- countAllelesPerYear(alleles = nwk$V$allele, sam = sam, mapping = assoc$mapping,
                            apam = assoc$alleles$A, nul = -30)

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
