% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__networkAnalyser__countGeneClassLinks.R
\name{countGeneClassLinks}
\alias{countGeneClassLinks}
\title{Count the number of links in a network at the level of gene classes}
\usage{
countGeneClassLinks(net, mapping, out.matrix = FALSE)
}
\arguments{
\item{net}{A Graph object, or a data frame whose first two columns are names
of allele Y (response variable) and X (explanatory variable).}

\item{mapping}{A data frame mapping allele names to gene classes. It
can be the data frame "mapping" in the output list of findPhysLink.}

\item{out.matrix}{A logical argument determining whether a square matrix or a
data frame is returned.}
}
\value{
A data frame of three columns: Class_y and Class_x, the target and
source gene class, respectively; Links, link counts per kind of edge (Class_y,
Class_x).
}
\description{
This function is useful in summarising connections between
gene classes (for example, antimicrobial classes) based on a network.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
