% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__networkAnalyser__countNeighbours.R
\name{countNeighbours}
\alias{countNeighbours}
\title{Count neighbours per node}
\usage{
countNeighbours(net, groups = NULL, na = "Mix")
}
\arguments{
\item{net}{A Graph object, which can be created by the function mkNetwork.}

\item{groups}{A data frame mapping alleles to groups, such as genes. The first
column stores allele names and the second column stores group information.}

\item{na}{A character argument specifying the default group in the output.}
}
\value{
When groups != NULL, the function returns a list of two data frames
for allele-level and gene-level neighbour counts, respectively. Otherwise, a
single data frame of allele-level neighbour counts is returned.
}
\description{
Count how many neighbours does each node have in a network. The
function treats the network as undirected.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
