% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__visualisation__drawHeatMap.R
\name{drawHeatMap}
\alias{drawHeatMap}
\title{Draw a heat map for a given variable}
\usage{
drawHeatMap(data = NULL, x = "x", y = "y", val = "val", diag = 1,
  replace.na = 0, cluster.row = TRUE, cluster.col = TRUE,
  cluster.method = "complete", dist.row = "euclidean",
  dist.col = "euclidean", colour.low = "white", colour.mid = NULL,
  colour.high = "red", colour.grad = 50, colour.breaks,
  display.val = FALSE, font.size = 10, filename = "heatmap.png",
  res = 72, width = 800, height = 800, unit = "px")
}
\arguments{
\item{data}{A data frame of at least three columns, such as x, y and z, or a matrix to be plotted directly as a heat map.}

\item{x}{Name (character) of the first variable of IDs.}

\item{y}{Name (character) of the second variable of IDs.}

\item{val}{Name (character) or index (integer) of the variable for values in the heat matrix.}

\item{diag}{The value of diagonal cells in the unclustered heat map.}

\item{replace.na}{The value used to replace NA's.}

\item{cluster.row}{Whether cluster rows hierarchically.}

\item{cluster.col}{Whether cluster columns hierarchically.}

\item{cluster.method}{The method for hierarchical clustering.}

\item{dist.row}{The distance metric for hierarchical clustering of rows.}

\item{dist.col}{The distance metric for hierarchical clustering of columns.}

\item{colour.low}{The colour for the lowest value in the heat matrix.}

\item{colour.mid}{The colour for the middle (that is, the "zero") value. Set to NULL if only a single colour transition are to be included.}

\item{colour.high}{The colour for the highest value.}

\item{colour.grad}{The number of colour grades.}

\item{colour.breaks}{Break points (colour.grad + 1) of values for assigning colours.}

\item{display.val}{A boolean value determining whether to overlay values on the heat map or not.}

\item{font.size}{Font size of labels in the heat map.}

\item{filename}{Path and name for the output PNG image.}

\item{res}{Resolution of the output figure. Default: 72 ppi.}

\item{width}{Width of the output image.}

\item{height}{Height of the output image.}

\item{unit}{The unit of the width and height of the output image. Valid values: "mm" and "px" (default).}
}
\description{
This is a generic function for plotting a heat map given two ID variables and a value variable.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
