% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__distanceAnalyser__evalPL.R
\name{evalPL}
\alias{evalPL}
\title{Evaluate evidence of physical linkage for positive associations with summary statistics of distance measurements}
\usage{
evalPL(
  lmms,
  min.beta = 0,
  max.p = 0.05,
  max.range = 2000,
  min.pIBD = 0.9,
  score.dist = FALSE
)
}
\arguments{
\item{lmms}{This argument can be a list produced by the function summariseDist,
which contains LMM parameters and distance measurements. It can also be the
data frame assoc in the output of findPhysLink or lmms in the output of function
lmm. When the latter two kinds of input are used, users can rescore the evidence
using new criteria.}

\item{min.beta}{only associations with beta's >= min.beta will be considered
as showing evidence of physical linkage.}

\item{max.p}{only associations with P values <= max.p will be considered as signficant.}

\item{max.range}{An integer specifying the maximum range (in bp) of in-group
allelic physical distances allowed to call consistent.}

\item{min.pIBD}{(optional) Minimum probability of the root of a minimum
inclusive clade to display a positive binary trait, such as having a pair of
alleles co-occurring or a specific allelic physical distance. Default: 0.9 (90\%).}

\item{score.dist}{(optional) A logical parameter specifying whether allelic
physical distances should be taken into account for scoring edges. Default: FALSE.}
}
\description{
This function evaluates evidence of physical linkage using association
status and consistency in allelic physical distances.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
