% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__extractPairedRows.R
\name{extractPairedRows}
\alias{extractPairedRows}
\title{Extract paired and unpaired rows from a data frame.}
\usage{
extractPairedRows(x, pair.col = "pair")
}
\arguments{
\item{x}{A data frame containing pair IDs.}

\item{pair.col}{A character or integer specifying which column is providing pair IDs.}
}
\description{
Assuming a data frame with a column storing pair IDs, this function extracts rows sharing the same pair
IDs from the data frame. It can be used to isolate symmetric associations from asymmetric associations.
Specifically, a pair of random variables X and Y is considered as in a symmetric association if both
models Y ~ X and X ~ Y display a significant effect of the explanatory variable, respectively. On the
contrary, X and Y is in an asymmetric association if either model shows insignificance.
}
\examples{
a <- rbind.data.frame(subset(assoc[["assoc"]], score >= 1.5 | (s_d == 1 & dif == 0)), stringsAsFactors = FALSE)
ap <- extractPairedRows(x = a, pair.id = "pair")

}
\author{
Yu Wan (\email{wanyuac@126.com})
}
