% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__findMinIncCladeOfStrains.R
\name{findMinIncCladeOfStrains}
\alias{findMinIncCladeOfStrains}
\title{Finding out and summarising the minimal inclusive clade containing all given strains}
\usage{
findMinIncCladeOfStrains(strains = NULL, clade.pam = NULL,
  clade.sizes = NULL)
}
\arguments{
\item{strains}{a vector of strain names}

\item{clade.pam}{a presence/absence matrix of samples in each clade of an input tree. It can be obtained
using the function tree2Clades of phylix.}

\item{clade.sizes}{(optional) A named vector of integers for the number of samples in each clade. It can
be obtained from the element "sizes" in the outputs of the function tree2Clades. Optional.}
}
\description{
This function works in a similar way as the function findMinIncClade, except it works on a
given set of strains rather than a pair of alleles.
}
\examples{
mic <- findMinIncCladeOfStrains(strains = c("strain1", "strain2"), clade.pam = assoc[["struc"]][["clades"]][["pam"]],
clade.sizes = assoc[["struc"]][["clades"]][["sizes"]])

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
