% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/func__geoTempAnalyser__getAllelesEarliestAppearance.R
\name{getAllelesEarliestAppearance}
\alias{getAllelesEarliestAppearance}
\title{Get earliest appearance of each allele in a vector}
\usage{
getAllelesEarliestAppearance(alleles, pam, sam)
}
\arguments{
\item{alleles}{A vector of allele names.}

\item{pam}{A presence-absence matrix of these alleles. It may come from the element
pam in the output list of the function ringPlotPAM.}

\item{sam}{A data frame of sample information. Required columns: Strain, Country, Year_low and Year_up.}
}
\description{
For a group of alleles, this function returns the earliest
occurrence time and country of each allele.
}
\examples{
emg_1 <- getAllelesEarliestAppearance(alleles = targets, pam = rp_1$pam, sam = sam)

}
\author{
Yu Wan (\email{wanyuac@126.com})
}
