% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__getAllelesPerPattern.R
\name{getAllelesPerPattern}
\alias{getAllelesPerPattern}
\title{Converting an allele-pattern data frame into a pattern-alleles data frame}
\usage{
getAllelesPerPattern(mapping, sep = ",")
}
\arguments{
\item{mapping}{the data frame alle.pat or mapping in outputs of the findPhysLink function. It can be other data frames, but must
contain names "allele" and "pattern".}

\item{sep}{the delimiter for alleles that belong to the same pattern in the output. It is a comma by default.}
}
\description{
This functions reformats the data frame "alle.pat" or mapping[, c("allele", "pattern")] in outputs of the findPhysLink function.
It is useful when a user wants to list all alleles under each particular pattern.

For example, given an input data frame:
    allele | pattern; a1 | 1; a2 | 1

The function returns a data frame containing a row:
    pattern | alleles; 1 | a1,a2
}
\examples{
assoc <- findPhysLink(...)
ap1 <- getAllelesPerPattern(assoc[["mapping"]], sep = ",")
ap2 <- getAllelesPerPattern(assoc[["alleles"]][["alle.pat"]], sep = ",")  # alternatively

}
\author{
Yu Wan (\email{wanyuac@126.com})
}
