% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__getAssocGenePair.R
\name{getAssocGenePair}
\alias{getAssocGenePair}
\title{Find out rows corresponding to alleles of a given pair of genes}
\usage{
getAssocGenePair(x, map, g1, g2, col1 = "y", col2 = "x")
}
\arguments{
\item{x}{A data frame to be subset for the result.}

\item{map}{A data frame to be searched against, which maps allele names to gene
information. There must be two columns named "allele" and "gene".}

\item{g1}{A string for the first gene name.}

\item{g2}{A string for the second gene name.}

\item{col1}{Name or index for the first column of alleles in x. Default: y.}

\item{col2}{Name or index for the second column of alleles in x. Default: x.}
}
\value{
A data frame from x.
}
\description{
This is a little tool function for getting all rows relating to
all alleles of two given genes.
}
\examples{
getAssocGenePair(x = assoc$mapping, g1 = "StrA_AGly", g2 = "DfrA8_Tmt", col1 = "y", col2 = "x")

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
