% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__getGeneClasses.R
\name{getGeneClass}
\alias{getGeneClass}
\title{Extracting gene classes from SRST2-formatted allele IDs.}
\usage{
getGeneClass(ids)
}
\arguments{
\item{ids}{A character vector of allele IDs.}
}
\description{
SRST2 defines each allele ID using several fields:
[sequence cluster ID]__[cluster symbol (the name of a gene, product class or group)]__[gene or allele name]__[sequence ID] [additional annotations separated by ";"].
For instance, a valid sequence ID is:
225__SulI_Sul__SulI__1616 no;no;SulI;Sul;U37105;4069-4908;840.
Usually, the cluster symbol is the antimicrobial class for an antimicrobial
resistance gene. This function aims to extract this symbol from the allele ID.
It is a simple function, however, this package includes it because of its
frequent usage.
}
\examples{
assoc <- lmm(...)
arg_cls <- data.frame(allele = assoc$mapping$allele, class = getGeneClass(x = assoc$mapping$gene), stringsAsFactors = FALSE)

}
\author{
Yu Wan (\email{wanyuac@126.com})
}
