% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__getRowsXY.R
\name{getRowsXY}
\alias{getRowsXY}
\title{Extract rows matching to a pair of keys k1 and k2.}
\usage{
getRowsXY(df, k1 = NULL, k2 = NULL, c1 = "y", c2 = "x")
}
\arguments{
\item{df}{a data frame whose rows are to be extracted.}

\item{k1}{a character, numeric or integer type key being searched against the column c1 (see below).}

\item{k2}{the same kind of key being searched against the column c2 (see below).}

\item{c1}{name or index of the first value column (variable) in df. Default: "y".}

\item{c2}{name or index of the second key column in df. Default: "x".}
}
\description{
This is a handy small tool to help investigating connections between two objects. A typical usage is to
extract or display results relating to alleles x and y.
}
\examples{
assoc <- findPhysLink(...)
z <- getRowsXY(df = assoc[["assoc"]], k1 = "IntI1_1.1", k2 = "SulI_1616", c1 = "y", c2 = "x")
z <- getRowsXY(df = assoc[["assoc"]], k1 = "IntI1_1.1", k2 = "SulI_1616", c1 = 1, c2 = 2)

}
\author{
Yu Wan (\email{wanyuac@126.com})
}
