% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__importAllelicPAM.R
\name{importAllelicPAM}
\alias{importAllelicPAM}
\title{Import the allelic presence-absence matrix (PAM) of bacterial genes.}
\usage{
importAllelicPAM(
  pam,
  pam.delim = "\\t",
  outliers = NULL,
  min.count = 1,
  alleles.inc = NULL,
  output.y = "",
  sample.order = NULL,
  skip = TRUE
)
}
\arguments{
\item{pam}{either the final PAM or the file name of a raw PAM created by the PAMmaker pipeline.
First column: Sample; other columns: allele identifiers.}

\item{pam.delim}{a single character for the delimiter in the text file of the allelic PAM.}

\item{outliers}{a vector of isolate names to be excluded from rows of the allelic PAM.}

\item{min.count}{the minimal number of times that an allele occurs in all isolates excluding outliers.
By default, only alleles that do not occur at all are removed (min.count = 1).
When min.count = 2, isolate-specific alleles are removed as well.}

\item{alleles.inc}{a vector of allele names to be included for columns of the allelic PAM.}

\item{output.y}{path to the output file for the centred pattern matrix Y.
Keep output.y = "" to prevent the function from printing the matrix.
In this case, only a list is returned to the user.}

\item{sample.order}{a vector of isolate names for the PAM to be matched with.
It can be used to filter the allelic PAM for in-group isolates.}

\item{skip}{wheter to skip overwriting existant output files.}
}
\description{
This function reads the allelic p/a matrix and returns a list of relevant matrices.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
