% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__distanceAnalyser__importPhysicalDists.R
\name{importPhysicalDists}
\alias{importPhysicalDists}
\title{Importing a table of physical distances as a data frame}
\usage{
importPhysicalDists(dists = NULL, delim = "\\t", ingroup = NULL,
  outgroup = NULL)
}
\arguments{
\item{dists}{The distance table to be imported.}

\item{delim}{A character specifying the delimiter in the distance table. Default:
a tab character.}

\item{ingroup}{A vector of strain names to be included in the result.}

\item{outgroup}{A vector of strains whose physical distances will be excluded
from the result.}
}
\value{
A data frame of physical distances.
}
\description{
This function reads a tabulated Bandage output into R for the
evaluation of distance consistency. The distance table can be created using
a helper pipeline physDist (github.com/wanyuac/physDist).
}
\author{
Yu Wan, \email{wanyuac@126.com}
}
