% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__networkAnalyser__mkNetwork.R
\name{mkNetwork}
\alias{mkNetwork}
\title{Create network data from association results.}
\usage{
mkNetwork(assoc = NULL, name.x = "x", name.y = "y", edge.attr = NULL,
  node.x.attr = NULL, node.y.attr = NULL, node.attr.names = "name",
  id = "")
}
\arguments{
\item{assoc}{A data frame from which the network is created. It may be pre-filtered
for some criteria by users.}

\item{name.x}{Column name (character) for the explanatory variable X.}

\item{name.y}{Column name (character) for the response variable Y.}

\item{edge.attr}{A character vector of column names for edge attributes.}

\item{node.x.attr}{A character vector of column names for node attributes of X.}

\item{node.y.attr}{A character vector of column names for node attributes of Y.
Vectors node.x.attr and node.y.attr must have the same length.}

\item{node.attr.names}{A character vector for column names of the final data frame
of nodes. It should have one more name (the "name" column for node names) than
do the vectors node.x.attr and node.y.attr.}

\item{id}{(optional) Network name}
}
\value{
G A Graph object having three slots: id, E and V. E: a data frame for
edges and their attributes; V, a data frame for nodes and their attributes; id,
an optional parameter for the network name.
}
\description{
This function generates edge and node tables for a network. A user
may print its output as an edge file and a node file for network construction.
}
\examples{
assoc <- findPhysLink(...)
nwk <- mkNetwork(assoc = assoc[["assoc"]], name.x = "x", name.y = "y",
edge.attr = c("beta", "score", "m", "s_d"), node.y.attr = "n_y",
node.x.attr = "n_x", node.attr.names = c("allele", "n"), id = "linkage")

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
