% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__retrieveAlleleSetInfo.R
\name{retrieveAlleleSetInfo}
\alias{retrieveAlleleSetInfo}
\title{Retrive allelic presence-absence information given a vector of allele names}
\usage{
retrieveAlleleSetInfo(targets, assoc, allelic.pam, min.count = 2,
  phylo.dists = NULL, dist.method = "binary",
  clust.method = "single")
}
\arguments{
\item{targets}{A character vector of allele names whose co-occurrence information is to be retrieved}

\item{assoc}{The element "assoc" in the output of findPhysLink.}

\item{allelic.pam}{The element "assoc$A" in the output of findPhysLink.}

\item{min.count}{(>= 1) The minimal number of alleles that an isolate must carry to be included in the extracted information}

\item{phylo.dists}{A matrix of phylogenetic distances between tips of a tree. It can be generated using the function cophenetic.phylo(tree).}

\item{dist.method}{The parameter for the method argument of the base function dist}

\item{clust.method}{The parameter for the method argument of the base function hclust}
}
\description{
This function is to be run after the findPhysLink pipeline.
}
\examples{
retrieveAlleleSetInfo(targets = c("aadA2", "dfrA12", "sul1"), assoc = a[["assoc"]], allelic.pam = a[["alleles"]][["A"]], min.count = 2, phylo.dists = cophenetic.phylo(tr))

}
\author{
Yu Wan (\email{wanyuac@126.com})
}
