% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__screePlotPCs.R
\name{screePlotPCs}
\alias{screePlotPCs}
\title{Calculating proportion of genetic variation captured by each principal
component}
\usage{
screePlotPCs(ev, ev.unsorted = FALSE)
}
\arguments{
\item{ev}{A vector of eigenvalues, which can be obtained from the element
struc$C$ev in the output of the functions lmm or findPhysLink.}

\item{ev.unsorted}{A logical argument telling the function that eigenvalues
have been sorted in a descending order. Default: FALSE.}
}
\description{
Since eigenvalues is proportional to the amount of genetic variation
captured by the corresponding principal component (PC), this function computes
the proportion for every PC so that users can draw a scree plot for PCs.
}
\examples{
pr <- screePlotPCs(ev = assoc_lmm$struc$C$ev)

}
\author{
Yu Wan (\email{wanyuac@126.com})
}
