% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__visualisation__showGeneContent.R
\name{showGeneContent}
\alias{showGeneContent}
\title{Draw a bubble plot and two bar plots to summarise gene and allele frequencies.}
\usage{
showGeneContent(af, gf, pam.g, d.min = 2, d.max = 20, fill.colour,
  border.colour = "grey90", core.genes = NULL, bar.acc = FALSE,
  panel.names = c("a", "b", "c"), panelA.xmax = 120,
  panelB.xinterv = 5, panelB.yinterv = 20, panelB.col = "grey50",
  panelC.lwd = 1, panelC.xinterv = 20, prev.out = NULL,
  f = "gene_content.png", w = 170, h = 200, r = 300, u = "mm",
  img.oma = c(0.1, 0.1, 0.1, 0.1), img.mar = c(3, 2.6, 1, 0.5),
  img.mgp = c(1.8, 0.6, -0.2))
}
\arguments{
\item{af}{A table of allele frequencies, produced by the function countAlleles.}

\item{gf}{A table of gene frequencies, produced by the function calcGeneFreq.}

\item{d.min}{Diameter for genes of the minimum allele count.}

\item{d.max}{Diameter for genes of the largest allele count.}

\item{fill.colour}{A named vector of colours filling bubbles.}

\item{border.colour}{A single colour for borders of circles.}

\item{core.genes}{Names of core genes}

\item{bar.acc}{Whether to draw a bar plot of gene frequencies for accessory genes
(TRUE) or all genes (FALSE, default).}

\item{panel.names}{A character vector of three elements for panel names.
Default: a, b, c.}

\item{panelA.xmax}{Maximum on the X axis of panel a for gene frequencies.
Default: 120.}

\item{panelB.xinterv}{Distance between major ticks on the X axis of the first
bar plot (panel b). Default: 5.}

\item{panelB.yinterv}{Distance between major ticks on the Y axis of the first
bar plot (panel b). Default: 20.}

\item{panelB.col}{Colour for a bar plot in the panel b. Default: grey50.}

\item{panelC.lwd}{An integer specifying the width of bars for allele frequencies
in the panel c. Default: 1.}

\item{panelC.xinterv}{Distance between major ticks on the X axis of the second
bar plot (panel b). Default: 50.}

\item{prev.out}{Output from a previous run of this function.}

\item{f}{Name for the output PNG file. Default: gene_content.png.}

\item{w}{Image width.}

\item{h}{Image height.}

\item{r}{Image resolution.}

\item{u}{Unit for the image width and height.}

\item{img.oma}{The argument oma for the function par during plotting.}

\item{img.mar}{The argument mar for the function par.}

\item{img.mgp}{The argument mgp for the function par.}
}
\description{
Use the function countAlleles and calcGeneFreq to know allele and
gene content before using this function, in order to set appropriate graphic
arguments. The diameter of each bubble is porportional to the allele number.

Dependency: ggplot2.
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
