% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__networkAnalyser__summariseCliques.R
\name{summariseCliques}
\alias{summariseCliques}
\title{Summarise allele content per clique}
\usage{
summariseCliques(q, s = NULL)
}
\arguments{
\item{q}{A list produced by the function max_cliques of the package igraph.}

\item{s}{An optional list produced by the function compEdgeOccur for edge similarities.}
}
\value{
A data frame of three columns: Clique, Size and Alleles.
}
\description{
This function extract allele names from cliques and return a
summary data frame.
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
