% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__geoTempAnalyser__tempNet.R
\name{tempNet}
\alias{tempNet}
\title{Creating a temporal network from graphs}
\usage{
tempNet(gs, tail_col = "node1", head_col = "node2", directed = FALSE,
  t_gap = 1, t_min = 1, t_max = 10, v_size = NULL, v_col = NULL,
  e_width = NULL, e_col = NULL, default_va = list())
}
\arguments{
\item{gs}{A GraphSet object with years as graph IDs.}

\item{tail_col}{Column name for tail vertices.}

\item{head_col}{Column name for head vertices.}

\item{directed}{A logical parameter specifying whether the network is directed
or not. Default: FALSE (undirected).}

\item{t_gap}{An integer or numeric argument determining the non-inclusive
minimum difference (namely, t > t_gap rather than t >= t_gap) between two
consecutive time points to be considered as not continuous.}

\item{v_col}{A single or multiple names for the columns from which temporal
vertex attributes are drawn and incorporated into the network. However, since
this function reserves "label" as an attribute, this name must not present in
this argument. Furthermore, va_col must not point to the first column, which
is used for vertex labels.}

\item{e_width}{Non-inclusive minimum edge weight that determines whether an edge
is present at a time point.}

\item{e_col}{Column name for edge weights.}

\item{default_va}{A named list of default attribute values for va_col.}
}
\description{
This function creates a temporal network for visualisation from
a list of Graph objects. Graphs in this list are named by years. Every member
graph must consist of the same vertices and edges. In particular, when the
network is undirectional, node order of each edge must be the same in the data
frame defining edges.

Dependency: package network and networkDynamic. The package ndtv is required
to display the output dynamic network as an animation.
}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
