% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__tree2Clades.R
\name{tree2Clades}
\alias{tree2Clades}
\title{Decompose a tree into a presence/absence matrix of samples in each clade of the tree}
\usage{
tree2Clades(tr, sample.order)
}
\arguments{
\item{tr}{A phylo object of an input tree.}

\item{sample.order}{A character vector of sample names to reorder tips of the input tree. In practice, the names
follow the order in the SNP matrix, the allelic PAM and the projection matrix C.}
}
\description{
This function is an adaptation of the tree2patterns function in the BugWAS package for phylix.
}
\examples{
assoc <- findPhysLink(...)
clade.pam <- treeToClades(tr = assoc[["C"]][["tr"]], sample.order = rownames(assoc[["C"]][["C"]]))

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
