% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__networkAnalyser__vertexAttr2Size.R
\name{vertexAttr2Size}
\alias{vertexAttr2Size}
\title{Mapping a vector of numerics to vertex sizes through a linear transformation}
\usage{
vertexAttr2Size(x, x.min = min(x), x.max = max(x), size.min = 15,
  size.max = 60)
}
\arguments{
\item{x}{Input numeric vector to be transformed.}

\item{x.min}{Minimum of x or the variable from which the x is drawn.}

\item{x.max}{Maximum of x or the variable from which the x is drawn.}

\item{size.min}{Minimal vertex size.}

\item{size.max}{Maximal vertex size.}
}
\value{
A vector of sizes with a precision of two decimals each.
}
\description{
This function generates vertex sizes for plotting networks based on a vertex attribute.
It works in the same way as the Cytoscape in mapping continuous values to an attribute. This is a
"smoother" way than simply multiply values with a constant.
}
\examples{
library(igraph)
...  # network construction
assoc <- findPhysLink(...)
V(net)$size <- vertexAttr2Size(x = V(net)$n, x.min = 2, x.max = max(assoc[["mapping"]]$count),
size.min = 15, size.max = 60)

}
\author{
Yu Wan (\email{wanyuac@gmail.com})
}
