% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__visualisation__heatMapPAM.R
\name{heatMapPAM}
\alias{heatMapPAM}
\title{heatMapPAM}
\usage{
heatMapPAM(p, data, col_colours = "black", null_colour = "grey90",
  border_colour = "white", cluster_cols = FALSE,
  cluster_method = "single", cluster_distance = "binary", colnames = TRUE,
  colnames_position = "bottom", colnames_angle = 0, colnames_level = NULL,
  set_label_colours = FALSE, colnames_offset_x = 0, colnames_offset_y = 0,
  font.size = 4, hjust = 0.5, offset = 0, width = 1,
  show_legend = FALSE)
}
\arguments{
\item{p}{A tree view from the function ggtree in the package ggtree}

\item{data}{A binary presence-absence matrix (PAM)}

\item{col_colours}{Either a vector of colour codes named by column names of
the PAM, or a single colour code for all columns.}

\item{null_colour}{The colour for absence status.}

\item{border_colour}{Colour of heatmap cell border}

\item{cluster_cols}{A logical argument determining if columns of PAM will be
clustered or not. Default: FALSE. Clustering will not run when there are less
than three columns in the PAM.}

\item{cluster_method}{Method for clustering. Default: binary.}

\item{cluster_distance}{The distance used for clustering columns. Default: binary.}

\item{colnames}{Logical, add matrix colnames or not}

\item{colnames_position}{One of 'bottom' or 'top'}

\item{colnames_angle}{Angle of column names}

\item{colnames_level}{Levels of colnames}

\item{set_label_colours}{Whether colour column names in the same way as painting columns.
By default, this option is turned off, which result in all column names to be printed
in black.}

\item{colnames_offset_x}{x offset for column names}

\item{colnames_offset_y}{y offset for column names}

\item{font.size}{Font size of matrix colnames}

\item{hjust}{hjust for column names (0: align left, 0.5: align center, 1: align righ)}

\item{offset}{Offset of heatmap to tree}

\item{width}{Total width of heatmap, compare to width of tree}

\item{show_legend}{A logical argument determining whether to show the legend.}
}
\value{
A list of three elements: p, a tree view that can be plotted directly;
mapping: a data frame mapping allele names to colours; data, a data frame of
weighted presence-absence status.
}
\description{
To append a heatmap of a presence-absence matrix (PAM) to the
right side of a phylogenetic tree. This function is modified from the function
gheatmap in the R package ggtree (github.com/GuangchuangYu/ggtree) through
replacing scale_fill_gradient with scale_fill_discrete. As a result, columns
in the heat map can be coloured by group information. This function is useful
when users need draw a heat map of gene or allele contents across strains.

In the PAM, absence status is coded as zero and presence status is coded as one.
}
\examples{
htmap <- heatMapPAM(...)
plot(htmap$p)

}
\author{
Guangchuang Yu, Yu Wan (\email{wanyuac@gmail.com})
}
