% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__geoTempAnalyser__mkCoocurNetwork.R
\name{mkCoocurNetwork}
\alias{mkCoocurNetwork}
\title{Converting a given network into a co-occurrence network}
\usage{
mkCoocurNetwork(net, pam = NULL, sam = NULL, years = 0, ds = NULL)
}
\arguments{
\item{net}{A Graph object defining a network whose edge attributes must include
"d_min" and "d_max" for the ingroup distances when the parameter d is specified.
The first two columns of the edge attribute must be node names.}

\item{pam}{An allelic presence-absence matrix produced by the function
importAllelicPAM.}

\item{sam}{A data frame of sample information. It must have two columns named
"Year_low" and "Year_up" for the lower bound and upper bound of years that an
isolate was obtained. For isolates whose isolation years are known explicitly,
Year_low = Year_up. In addition, the first column must be isolate names.}

\item{years}{An integer vector for every year the co-occurrence network is
generated. Leave this argument as 0 (default) to disable year-based sampling.}

\item{ds}{A data frame of allelic physical distances, which is generated by the
physDist pipeline.}
}
\value{
A list of two elements. The first element "G" is a list of Graph objects,
each named by a year; the second element w is a numeric vector for weighted
occurrence count per isolate. However, when years <= 0, the G element is a
single Graph object.
}
