% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func__geoTempAnalyser__tempNet.R
\name{tempNet}
\alias{tempNet}
\title{Creating a temporal network from graphs}
\usage{
tempNet(
  gs,
  v.label = "allele",
  e.tail = "node1",
  e.head = "node2",
  directed = FALSE,
  t.gap = 1,
  v.value = "count",
  v.value.base = 0,
  v.size.min = 1,
  v.size.max = 25,
  v.colour = "red",
  e.weight = "n_xy",
  e.weight.base = 1,
  e.weight.cutoff = 0,
  e.width.min = 1,
  e.width.max = 10,
  e.colour = "m",
  e.colour.low = "#FFA0A0",
  e.colour.high = "#FF0000",
  e.colour.num = 9
)
}
\arguments{
\item{gs}{A GraphSet object with years as graph IDs.}

\item{v.label}{A single name or index for the column from which vertex labels
are drawn.}

\item{e.tail}{Column name for tail vertices.}

\item{e.head}{Column name for head vertices.}

\item{directed}{A logical parameter specifying whether the network is directed
or not. Default: FALSE (undirected).}

\item{t.gap}{An integer or numeric argument determining the non-inclusive
minimum difference (namely, t > t.gap rather than t >= t.gap) between two
consecutive time points to be considered as not continuous. This arguments
determines each onset and terminus range for edges.}

\item{v.value}{A character or integer specifying the column in V(gs) used as
the variable defining vertex sizes.}

\item{v.value.base}{A numeric specifying the minimum value for the variable
(column) defining vertex sizes. For instance, you may set v.size.base for a
dynamic co-occurrence count where the expected minimum count is zero.}

\item{v.size.min}{An integer for the minimum vertex size (when v.size.base is
reached).}

\item{v.size.max}{An integer for the maximum vertex size.}

\item{v.colour}{A character vector of colours named by vertex labels.}

\item{e.weight}{Column name or index in E(gs[[t]]) for edge widths.}

\item{e.weight.base}{The minimum valid value of edge weights. It works in the
same way as v.value.base.}

\item{e.weight.cutoff}{Non-inclusive minimum edge weight that determines whether an edge
is present at a time point.}

\item{e.width.min}{Minimum edge width.}

\item{e.width.max}{Maximum edge width.}

\item{e.colour}{A column name or index for the attributed used for colour assignment.}

\item{e.colour.low}{A character specifying the colour for edges of the least weight.}

\item{e.colour.high}{A character specifying the colour for edges of the highest weight.}

\item{e.colour.num}{An integer specifying the number of colours to be assigned.}
}
\description{
This function creates a temporal network for visualisation from
a list of Graph objects. Graphs in this list are named by years. Every member
graph must consist of the same vertices and edges. In particular, when the
network is undirectional, node order of each edge must be the same in the data
frame defining edges.

Dependency: package network and networkDynamic. The package ndtv is required
to display the output dynamic network as an animation.
}
\author{
Yu Wan (\email{wanyuac@126.com})
}
